requirements.txt
# Web Framework
fastapi==0.133.0
uvicorn==0.41.0
streamlit==1.54.0

# PDF & Docx Parsers
pymupdf==1.27.1
python-docx==1.2.0
python-multipart==0.0.22

# Data Validation
pydantic==2.12.5

# Logging
loguru==0.7.3

# LLM Framework
langchain==1.2.10
langchain-openai==1.1.10
langchain-core==1.2.15

# Environment Variables
python-dotenv==1.2.1